requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
tqdm>=4.65.0
python-dotenv>=1.0.0
zeep>=4.2.0
lxml>=4.9.0
